+++ MiniBot/MiniBot/Products/Product.cs
@@ -7,14 +7,14 @@ namespace MiniBot.Products
 {
     abstract class Product : IProduct, IShowInfo
     {
-        public short Id { get; private set; }
+        public int Id { get; private set; }
         public float Cost { get; private set; }
         public string Name { get; private set; }
         public string Description { get; private set; }
         public byte Score { get; set; }
         public byte Discount { get; set; }
 
-        public Product(short id, string name, float cost, byte score, string description)
+        public Product(int id, string name, float cost, byte score, string description)
         {
             Id = id;
             Name = name;
